requirements.txt
Flask>=2.2.5
Authlib
Werkzeug==2.2.2
Flask-SQLAlchemy==3.1.1
openai
oauthlib==2.1.0
aiohttp==3.9.0b0
python-dotenv